+++ thoth-api/src/graphql/mod.rs
@@ -1,3 +1,6 @@
 #[cfg(feature = "backend")]
 pub mod model;
 pub mod utils;
+
+#[cfg(feature = "backend")]
+pub use juniper::http::GraphQLRequest;